fix: Resolver styles todo-header button center
